Fix code review separate,components,update Readme and fix console errors
