+++ src/containers/__test__/FilterCategory.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/extensions */
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { createStore } from 'redux';
